finished
i tested it and everything works. if you test it and that is not the
case please tell me. Christian, I only changed the account,
asthmacontroller, and createaccount pages. however you will be better
off copying and pasting the asthma controller with the amount of changes
i made. amber said my gui is the old one so just make sure the my names
line up with the new page