fixed error checking
I fixed the issues Anna mentioned last time. Also, I looked at that wrap
text option. It didn't work the way I wanted it to so I just moved the
labels over more. Amber can mess with it and design it better. I, also,
added some fx:id tags on the fxml page for tabpane and tabs as well as a
fair amount of changes to the aap error checking and a small change to
the asthmacontroller